# backend/storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fishbook API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/fishes": {
            "get": {
                "tags": [
                    "Fishes"
                ],
                "operationId": "listFishes",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "breed",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "breed",
                                "created_at",
                                "size"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedFishCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fishes"
                ],
                "operationId": "createFish",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFishRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FishResourceEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/fishes/{fish}": {
            "get": {
                "tags": [
                    "Fishes"
                ],
                "operationId": "getFish",
                "parameters": [
                    {
                        "name": "fish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid",
                            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FishResourceEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fishes"
                ],
                "operationId": "deleteFish",
                "parameters": [
                    {
                        "name": "fish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid",
                            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Fishes"
                ],
                "operationId": "updateFish",
                "parameters": [
                    {
                        "name": "fish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid",
                            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFishRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FishResourceEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/repos/{owner}/{repo}/aquarium": {
            "get": {
                "tags": [
                    "Repo Aquarium"
                ],
                "operationId": "getRepoAquarium",
                "parameters": [
                    {
                        "name": "owner",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepoAquariumResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepoAquariumErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepoAquariumErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/repos/{owner}/{repo}/fork-to-my-aquarium": {
            "post": {
                "tags": [
                    "Repo Aquarium"
                ],
                "operationId": "forkRepoAquarium",
                "parameters": [
                    {
                        "name": "owner",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RepoForkResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreFishRequest": {
                "required": [
                    "nickname",
                    "breed",
                    "color_hex",
                    "size"
                ],
                "properties": {
                    "nickname": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1
                    },
                    "breed": {
                        "type": "string",
                        "example": "guppy"
                    },
                    "color_hex": {
                        "type": "string",
                        "pattern": "^#[0-9A-Fa-f]{6}$",
                        "example": "#FF6B9D"
                    },
                    "size": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "UpdateFishRequest": {
                "properties": {
                    "nickname": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1
                    },
                    "color_hex": {
                        "type": "string",
                        "pattern": "^#[0-9A-Fa-f]{6}$"
                    },
                    "size": {
                        "type": "integer",
                        "maximum": 100,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
...
            "FishResource": {
                "required": [
                    "id",
                    "nickname",
                    "breed",
                    "color_hex",
                    "size",
                    "source",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "ulid",
                        "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$",
                        "example": "01HG5W7DZ8KX9F3N2VYRMPQABT"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "breed": {
                        "type": "string"
                    },
                    "color_hex": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "manual",
                            "github_repo"
                        ]
                    },
                    "source_ref": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "RepoStats": {
                "properties": {
                    "stars": {
                        "type": "integer"
                    },
                    "forks": {
                        "type": "integer"
                    },
                    "issues": {
                        "type": "integer"
                    },
                    "watchers": {
                        "type": "integer"
                    },
                    "contributors": {
                        "type": "integer"
                    },
                    "language": {
                        "type": "string",
                        "nullable": true
                    },
                    "age_days": {
                        "type": "integer"
                    },
                    "fetched_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "RepoFishItem": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "breed": {
                        "type": "string"
                    },
                    "color_hex": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "source_ref": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RepoAquariumResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "stats": {
                                "$ref": "#/components/schemas/RepoStats"
                            },
                            "fish_set": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RepoFishItem"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "RepoAquariumErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RepoForkResponse": {
                "properties": {
                    "added": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "PaginatedFishCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FishResource"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "FishResourceEnvelope": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/FishResource"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}